fastapi-app/requirements.txt
anyio==3.6.1
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.26.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
distlib==0.3.4
executing==0.8.3
fastapi==0.85.0
filelock==3.4.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
jedi==0.18.1
logging-tree==1.9
logx==0.1.1579232358
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.1
Pygments==2.11.2
pyodbc==4.0.34
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
stack-data==0.2.0
starlette==0.20.4
traitlets==5.1.1
typing_extensions==4.3.0
uvicorn==0.18.3
virtualenv==20.10.0
watchfiles==0.16.1
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.3
